Add a check for whether data should be shuffled
The test datasets should not be shuffled so that (i) the labels do not
have to be saved alongside the data every time and (ii) the change in
position could be reconstructed (not that this is planned anytime
soon).

The full data without transformations is not shuffled as of now, since
it will be used to generate the features representation later on.