Change CE Loss slightly
Use both augmentations for the CE Loss.

This should make the training more efficient since I am using the data
more efficiently.  But I am not sure whether this has any influence on
overfitting to the samples in the batch; since they are now seen
twice.  But at least this is the case for all samples.