Change elapsed_time to float/seconds
This change saves space on the disk and makes it easier to interpret
the raw numbers.  The resolution should still be fine since the
measure is noisy anyways.  Float16 will also only use a quarter of the
space of an int64.

Oddly, it seems that all of the elapsed times inside of batch do not
add up to t_batch.  Maybe this will also be fixed by that change.  Not
sure why that is happening in the first place, overflow should not be
happening with the dtype that I was using.