Fix torch seed in model creation and training
This will always keep the seed fixed in those two files (since the
seed for `rng` is fixed).  Thus the weights for the model are
reproducible, or rather dependant on the `random_state` passed to the
wrapper function.  The same holds for the training regime, which also
makes the iteration order of the dataloader reproducible (as far as I
understand it, unless there are inherent data races in the DataLoader
code itself).