Fix random_seed passing and refactor .part.csv
Now it also uses names.py to form the strings for redo.  This then
makes the caching easier, since all the other programs use it as well.
It's a tiny bit unfortunate that I specify infonce one too many times
when finetuning (when I already trained with euclidean metric), but
that is a minor things that it not worth the hassle of changing now.