Fix callback saving issues
An error occured on the mlcloud where despite calling `unlink()` just
before linking the file the target exists, which then raised an error
in `os.link()`.  This way should actually be more robust and replace
the (possibly existing) file.  Rather annoyed that issues like that
can creep up on such a big NFS.