Add the rest of the code for arch figure
Add the resnet and projection head blocks in the middle.

Un-normalize the transformed images so that they plot normally.

Add labels to the arrows.

Add plot of losses.

Change seed so that we get a different image sample and transformation.

Changing the order of \mathbf and \mathrm actually makes them rm and
bold.