Implement most of Dmitrys suggestions
Also remove loading of the unused model

Dmitrys suggestion in full:
panel B: I think the embedding can be made larger, so that it reaches the top of the image and also almost reaches panel C below.
panel B: maybe write "CIFAR-10" somewhere in the upper-left corner
panel C: the font size for "linear layer finetuning" is too small (when printed). same for axis tick labels. I think the smallest font size should be as in the "finetuning in 2D".
panel C: I would remove the word "default" because it's not really default as we are using Euclidean loss
panel C: vertical lines completely disappear when printed. they should be bolder (thicker and/or darker color)
panel A: "data augmentation" can be written in smaller font
panel A: I would remove labels "ResNet(.)" and "proj.head(hi)" from the arrows, I think it's clear because these things are labeled in the drawing, so this is repetitive
panel A: I had written make all h_i and z_i bold, but I think you already did in the last version :-)
panel A: make sure that in the formula the norm sign is \| and not ||
panel A: maybe replace ResNet label with ResNet18
panel A: add horizontal arrow that goes from ResNet to the 512 layer
panel A: I would remove the letter D from layer dimensionality because it looks a bit like 0 when printed in this small font
panel A: actually maybe write 512 / 1014 / 2 above the layers horizontally, and as we want to emphasize that the output is 2-dim, maybe make all three labels black color and larger font
panel A: "proj. head" -> "projection head"? there is space
between panels A and B: the lines disappear when printed, see point 5 above.
regarding the image, the car is okay, but can we choose the top augmentation to be more colorful? the flip is good, but greyscale looks a bit boring. My suggestion for the picture would be something really bright (more eye candy) and not having white background (so that it looks like a square)... Like an airborn plane in the blue sky... But I don't know, maybe the car is also fine.
panel A: maybe bring the upper branch of the graph a bit downwards and the lower branch a bit upwards, so that there is less vertical space between them and the neural net schematic.