Change arch.pdf to (hopefully) final layout
I changed the upper and lower row to use a var for the y-coord, so
it's easier to change.

Removed margins from scatter plot, now it is maximal.

Set the aspect of the loss plot so that there is enough white space
around it.

Fix padding and xticks in loss plot.

Shave off some more whitespace and reduce figheight.